subclass MakePathSafeAction from Highlander
